AR-611: make screen-columns.jsx
