+++ src/components/screen/screen-form.jsx
@@ -53,7 +53,7 @@ function ScreenForm({
   const { data: layouts } = useGetV1LayoutsQuery({
     page: 1,
     itemsPerPage: 20,
-    orderBy: "createdAt",
+    orderBy: "desc",
     order: "desc",
   });
